+++ app/src/main/java/com/example/mycontacts/ContactsApplication.kt
@@ -1,8 +1,6 @@
 package com.example.mycontacts
 
 import android.app.Application
-import com.example.mycontacts.room.AppContainer
-import com.example.mycontacts.room.AppDataContainer
 
 
 class ContactsApplication : Application() {
