--- app/src/main/java/com/example/mycontacts/contact/detail_composables/DetailsBodyScreen.kt
@@ -1,85 +1,0 @@
-package com.example.mycontacts.contact.detail_composables
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.OutlinedButton
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import com.example.mycontacts.R
-import com.example.mycontacts.contact.entry_composables.ContactInputForm
-import com.example.mycontacts.contact.ContactUiState
-
-@Composable
-fun DetailsBody(
-    contactUiState: ContactUiState,
-    onDelete: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var deleteConfirmationRequired by rememberSaveable { mutableStateOf(false) }
-
-    Column(
-        modifier = modifier
-            .padding(16.dp)
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        Box(modifier = modifier.align(alignment = Alignment.CenterHorizontally)) {
-            Image(
-                painter = painterResource(R.drawable.ic_profile),
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .size(100.dp)
-                    .clip(CircleShape)
-                    .border(
-                        width = 2.dp,
-                        color = MaterialTheme.colors.onPrimary,
-                        shape = CircleShape
-                    )
-            )
-        }
-
-        ContactInputForm(
-            contactUiState = contactUiState,
-            enabled = false,
-        )
-
-        OutlinedButton(
-            onClick = { deleteConfirmationRequired = true },
-            modifier = modifier
-                .fillMaxWidth()
-        ) {
-            Text(text = stringResource(R.string.delete_button))
-        }
-
-        if (deleteConfirmationRequired) {
-            DeleteConfirmationDialog(
-                onDeleteConfirm = {
-                    deleteConfirmationRequired = false
-                    onDelete()
-                },
-                onDeleteCancel = { deleteConfirmationRequired = false })
-        }
-    }
-}
