+++ app/src/main/java/com/example/mycontacts/domain/repository/ContactsRepository.kt
@@ -1,0 +1,17 @@
+package com.example.mycontacts.domain.repository
+
+import com.example.mycontacts.domain.model.Contact
+import kotlinx.coroutines.flow.Flow
+
+interface ContactsRepository {
+
+    fun getAllContactsStream(): Flow<List<Contact>>
+
+    fun getContactStream(id: Int): Flow<Contact?>
+
+    suspend fun insertContact(contact: Contact)
+
+    suspend fun deleteContact(contact: Contact)
+
+    suspend fun updateContact(contact: Contact)
+}
