--- app/src/main/java/com/example/mycontacts/room/Contact.kt
@@ -1,15 +1,0 @@
-package com.example.mycontacts.room
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "contacts")
-data class Contact(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val firstName: String,
-    val lastName: String,
-    val address: String,
-    val gender: String,
-)
-
