--- app/src/main/java/com/example/mycontacts/room/ContactsDatabase.kt
@@ -1,24 +1,0 @@
-package com.example.mycontacts.room
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-
-@Database(entities = [Contact::class], version = 1, exportSchema = false)
-abstract class ContactsDatabase : RoomDatabase() {
-    abstract fun contactDao(): ContactDao
-
-    companion object {
-        @Volatile
-        private var Instance: ContactsDatabase? = null
-
-        fun getDatabase(context: Context): ContactsDatabase {
-            return Instance ?: synchronized(this) {
-                Room.databaseBuilder(context, ContactsDatabase::class.java, "contacts_database")
-                    .build()
-                    .also { Instance = it }
-            }
-        }
-    }
-}
